+++ packages/client/src/lib/index.ts
@@ -3,3 +3,11 @@ export * from './types';
 export * from './client';
 export * from './struct/geometry';
 export { LogLevel, type LoggerModule, setLogLevel, setModuleLogLevel, getModuleLogLevel } from './util/logger';
+export {
+  type RobotPlatform,
+  type ParsedRobotAddress,
+  getRobotAddress,
+  getTeamIpAddress,
+  parseRobotAddress,
+  SYSTEMCORE_MDNS_HOST,
+} from './util/util';